Refactor pattern filter
